requirements.txt
django==4.2.9
djangorestframework==3.14.0
drf-spectacular==0.27.1
gunicorn
faker==22.6.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
transliterate==1.10.2